+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- A lot more information on the stats
 - Not requesting the signed and altered cards
 - Switch to a mandatory config path for strategies' options
 - Adding options to change the number of cards requested per language, the default max results